Removed the Context type and inlined all mappings
