+++ crates/rune-cli/src/build.rs
@@ -201,7 +201,7 @@ pub(crate) fn analyze(
 
     let parsed =
         Document::parse(&src).context("Unable to parse the Runefile")?;
-    let rune = hotg_rune_syntax::analyse(&parsed, &mut diags);
+    let rune = hotg_rune_syntax::analyse(parsed, &mut diags);
 
     let mut writer = StandardStream::stderr(color);
     let config = Config::default();
