+++ crates/syntax/src/passes/construct_pipeline.rs
@@ -1,11 +1,21 @@
 use codespan_reporting::diagnostic::Diagnostic;
 use indexmap::IndexMap;
 
-use crate::{hir::HirId, passes::Context, yaml::Stage};
+use crate::{
+    Diagnostics,
+    hir::{HirId, NameTable, Node, Slot},
+    yaml::Stage,
+};
 
-pub(crate) fn run(ctx: &mut Context<'_>, pipeline: &IndexMap<String, Stage>) {
+pub(crate) fn run(
+    pipeline: &IndexMap<String, Stage>,
+    names: &NameTable,
+    stages: &mut IndexMap<HirId, Node>,
+    slots: &mut IndexMap<HirId, Slot>,
+    diags: &mut Diagnostics,
+) {
     for (name, stage) in pipeline {
-        let node_id = match ctx.rune.get_id_by_name(name) {
+        let node_id = match names.get_id(name) {
             Some(id) => id,
             None => continue,
         };
@@ -13,20 +23,20 @@ pub(crate) fn run(ctx: &mut Context<'_>, pipeline: &IndexMap<String, Stage>) {
         let mut input_slots = Vec::new();
 
         for input in stage.inputs() {
-            let incoming_node_id = match ctx.rune.get_id_by_name(&input.name) {
+            let incoming_node_id = match names.get_id(&input.name) {
                 Some(id) => id,
                 None => {
                     let diag = Diagnostic::error().with_message(format!(
                         "No node associated with \"{}\"",
                         input
                     ));
-                    ctx.diags.push(diag);
+                    diags.push(diag);
                     input_slots.push(HirId::ERROR);
                     continue;
                 },
             };
 
-            let incoming_node = ctx.rune.get_stage(&incoming_node_id).unwrap();
+            let incoming_node = stages.get(&incoming_node_id).unwrap();
 
             if incoming_node.output_slots.is_empty() {
                 let diag = Diagnostic::error().with_message(format!(
@@ -34,7 +44,7 @@ pub(crate) fn run(ctx: &mut Context<'_>, pipeline: &IndexMap<String, Stage>) {
                             name,
                             input
                         ));
-                ctx.diags.push(diag);
+                diags.push(diag);
                 input_slots.push(HirId::ERROR);
                 continue;
             }
@@ -43,7 +53,7 @@ pub(crate) fn run(ctx: &mut Context<'_>, pipeline: &IndexMap<String, Stage>) {
             match incoming_node.output_slots.get(input_index).copied() {
                 Some(slot_id) => {
                     input_slots.push(slot_id);
-                    let slot = ctx.rune.slots.get_mut(&slot_id).unwrap();
+                    let slot = slots.get_mut(&slot_id).unwrap();
                     slot.output_node = node_id;
                 },
                 None => {
@@ -53,14 +63,14 @@ pub(crate) fn run(ctx: &mut Context<'_>, pipeline: &IndexMap<String, Stage>) {
                             input,
                             incoming_node.output_slots.len(),
                         ));
-                    ctx.diags.push(diag);
+                    diags.push(diag);
                     input_slots.push(HirId::ERROR);
                     continue;
                 },
             }
         }
 
-        let node = ctx.rune.get_stage_mut(&node_id).unwrap();
+        let node = stages.get_mut(&node_id).unwrap();
         node.input_slots = input_slots;
     }
 }
@@ -69,8 +79,10 @@ pub(crate) fn run(ctx: &mut Context<'_>, pipeline: &IndexMap<String, Stage>) {
 mod tests {
     use crate::{
         Diagnostics,
-        passes::{construct_pipeline, register_output_slots, register_stages},
-        utils::dummy_document,
+        passes::{
+            construct_pipeline, helpers, register_output_slots, register_stages,
+        },
+        utils::{Builtins, HirIds, dummy_document},
         yaml::{Document, DocumentV1},
     };
 
@@ -82,10 +94,35 @@ mod tests {
             Document::V1(DocumentV1 { pipeline, .. }) => pipeline,
         };
         let mut diags = Diagnostics::new();
-        let mut ctx = Context::new(&mut diags);
-        register_stages::run(&mut ctx, &pipeline);
-        register_output_slots::run(&mut ctx, &pipeline);
-        run(&mut ctx, &pipeline);
+        let mut ids = HirIds::new();
+        let builtins = Builtins::new(&mut ids);
+        let mut names = NameTable::default();
+        let spans = IndexMap::default();
+        let mut stages = IndexMap::default();
+        let resources = IndexMap::default();
+        let mut types = IndexMap::default();
+        let mut slots = IndexMap::default();
+
+        register_stages::run(
+            &mut ids,
+            &pipeline,
+            &spans,
+            &mut stages,
+            &resources,
+            &mut names,
+            &mut diags,
+        );
+        register_output_slots::run(
+            &mut ids,
+            &pipeline,
+            &mut types,
+            &builtins,
+            &names,
+            &mut slots,
+            &mut stages,
+            &mut diags,
+        );
+        run(&pipeline, &names, &mut stages, &mut slots, &mut diags);
         let edges = vec![
             ("audio", "fft"),
             ("fft", "model"),
@@ -93,15 +130,21 @@ mod tests {
             ("label", "output"),
         ];
 
-        construct_pipeline::run(&mut ctx, &pipeline);
+        construct_pipeline::run(
+            &pipeline,
+            &names,
+            &mut stages,
+            &mut slots,
+            &mut diags,
+        );
 
-        assert!(ctx.diags.is_empty(), "{:?}", ctx.diags);
+        assert!(diags.is_empty(), "{:?}", diags);
         for (from, to) in edges {
             println!("{:?} => {:?}", from, to);
-            let from_id = ctx.rune.get_id_by_name(from).unwrap();
-            let to_id = ctx.rune.get_id_by_name(to).unwrap();
+            let from_id = names.get_id(from).unwrap();
+            let to_id = names.get_id(to).unwrap();
 
-            assert!(ctx.rune.has_connection(from_id, to_id));
+            assert!(helpers::has_connection(&stages, from_id, to_id));
         }
     }
 }
